refactor(auth)!: inject auth dependencies
Move authentication middleware from global functions to injected service
to eliminate hidden dependencies and improve testability.

BREAKING CHANGE: AuthMiddleware and OptionalAuthMiddleware are now
methods on UserAuth struct requiring dependency injection in handlers.

* Move user_handler/user_auth.go → auth/auth.go
* Create UserAuthInterface for abstraction
* Add MockUserAuth following testify/mock patterns
* Update TaskHandler/UserHandler constructors to accept auth dependency
* Wire UserAuth instance in main.go dependency injection
* Update all handler tests with auth mocks
* Move and update auth tests to auth package
