+++ internal/auth/auth.go
@@ -1,0 +1,126 @@
+package auth
+
+import (
+	"net/http"
+	"strconv"
+	"strings"
+	"taskflow/internal/common"
+	"taskflow/pkg"
+
+	"github.com/gin-gonic/gin"
+)
+
+type UserAuth struct {
+	secretKey []byte
+}
+
+func NewUserAuth(secret string) *UserAuth {
+	return &UserAuth{secretKey: []byte(secret)}
+}
+
+var _ UserAuthInterface = (*UserAuth)(nil)
+
+func (ua *UserAuth) AuthMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		secretKey := ua.secretKey
+
+		authHeader := c.GetHeader("Authorization")
+		if authHeader == "" {
+			c.JSON(http.StatusUnauthorized, common.ErrorResponse{
+				Message: "authorization header required",
+			})
+			c.Abort()
+			return
+		}
+
+		parts := strings.SplitN(authHeader, " ", 2)
+		if len(parts) != 2 || parts[0] != "Bearer" {
+			c.JSON(http.StatusUnauthorized, common.ErrorResponse{
+				Message: "invalid authorization header format",
+			})
+			c.Abort()
+			return
+		}
+
+		tokenString := parts[1]
+
+		claims, err := pkg.ValidateToken(tokenString, secretKey)
+		if err != nil {
+			c.JSON(http.StatusUnauthorized, common.ErrorResponse{
+				Message: "invalid or expired token",
+			})
+			c.Abort()
+			return
+		}
+
+		// Handle numeric or string user_id from claims
+		claimsMap := *claims
+		var userID int
+		switch v := claimsMap["user_id"].(type) {
+		case float64:
+			userID = int(v)
+		case int:
+			userID = v
+		case string:
+			var err error
+			userID, err = strconv.Atoi(v)
+			if err != nil {
+				c.JSON(http.StatusUnauthorized, common.ErrorResponse{
+					Message: "invalid user ID in token",
+				})
+				c.Abort()
+				return
+			}
+		default:
+			c.JSON(http.StatusUnauthorized, common.ErrorResponse{
+				Message: "invalid token claims",
+			})
+			c.Abort()
+			return
+		}
+
+		// Set userID in context for handlers
+		c.Set("userID", userID)
+		c.Next()
+	}
+}
+
+// OptionalAuthMiddleware allows routes to continue even if JWT is missing/invalid
+func (ua *UserAuth) OptionalAuthMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		secretKey := ua.secretKey
+
+		authHeader := c.GetHeader("Authorization")
+		if authHeader == "" {
+			c.Next()
+			return
+		}
+
+		parts := strings.SplitN(authHeader, " ", 2)
+		if len(parts) != 2 || parts[0] != "Bearer" {
+			c.Next()
+			return
+		}
+
+		tokenString := parts[1]
+		claims, err := pkg.ValidateToken(tokenString, secretKey)
+		if err != nil {
+			c.Next()
+			return
+		}
+
+		claimsMap := *claims
+		switch v := claimsMap["user_id"].(type) {
+		case float64:
+			c.Set("userID", int(v))
+		case int:
+			c.Set("userID", v)
+		case string:
+			if userID, err := strconv.Atoi(v); err == nil {
+				c.Set("userID", userID)
+			}
+		}
+
+		c.Next()
+	}
+}
