+++ internal/auth/auth_interface.go
@@ -1,0 +1,8 @@
+package auth
+
+import "github.com/gin-gonic/gin"
+
+type UserAuthInterface interface {
+	AuthMiddleware() gin.HandlerFunc
+	OptionalAuthMiddleware() gin.HandlerFunc
+}
